+++ Assets/Scenes/MainGame/Buyer.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b0a6fbd9a0f204b7e8c366b7b5185410
+folderAsset: yes
+timeCreated: 1512374630
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
