+++ Assets/Scenes/MainGame/Buyer/Buyer.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 24809080af4f645f98118cac4c1326f6
+timeCreated: 1512375006
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
