+++ Assets/Scenes/MainGame/Buyer/BuyerSpawner.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 70e4732d4ee684f87bd61a19589c8bd1
+timeCreated: 1512375249
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
