+++ Assets/Scenes/MainGame/Buyer/BuyerSpawner.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 78a57958425954e06bf4bf478140c827
+timeCreated: 1512375237
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
