--- Assets/Scenes/MainGame/Enemy/PolicemanSpawner.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 1b5ea86fae5ce4bb0a671c644864c049
-timeCreated: 1512367670
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
