+++ Assets/Scenes/MainGame/MainGame/LightingData.asset.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 583a85bf652a3462b809ecbc6a2cb8bc
+timeCreated: 1512376556
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 25800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
