+++ Assets/Scenes/MainGame/MainGame/NavMesh.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: 85844dc90c8b64b3194333f4c5b2cfa6
-timeCreated: 1512371264
+timeCreated: 1512376432
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
