+++ Assets/Scenes/MainGame/SpawnerSpiner.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SpawnerSpiner : MonoBehaviour {
+
+	// Use this for initialization
+	void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		transform.Rotate (new Vector3 (0f, 0.5f + Random.value * 0.5f, 0f));
+	}
+}
