+++ Assets/Scenes/MainGame/SpawnerSpiner.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e87b19174d22143dbbd8151fe7b7ad9d
+timeCreated: 1512375885
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
